requirements.txt
django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.5,<2.9.0
autopep8>=1.5.4,<1.6.0
flake8>=3.8.3,<3.9.0
requests>=2.24.0,<2.30.0
djangorestframework-jwt>=1.11.0,<1.12.0
Pillow>=2.2.1,<=2.3.0
django-cors-headers>= 3.5.0,<3.6.0
requests-mock>= 1.8.0,<2.0.0
django-extensions>= 3.0.8,<3.1.0
django-ipware>= 3.0.1,<3.1.0